Update Search to Use Dispatch
Update the Search component to use dispatch instead of getting the
search directly from the server so that the search can happen locally
since it more likely has the latest data version and gets updated all
the time. Furthermore, it reduces the time going over the wire and has
the search locally, which is faster.