Add Cache Orders to Control Order Search Tree
Add cache orders to control the Order search tree because there can only
be one cache running in the application. So a singleton class can run
the adding charges and allow the developer to access the Order search
tree to see what is inside it.

Since the Order search tree is not used for display, it is too slow to
convert it into an array that the UI can use. However, we can still use
it for searching and let it be where the full orders are located,
whereas the orders in the state can be used for display. Those we can
use to remove orders that have already been delivered.