Add Cache Orders Find to Search Reducer
Add Cache Order Find to Search Reducer so the reducer can access Order
Search Tree and find the orders the user is looking for since a tree
structure is faster to search for than searching in an unordered array.
The array will need to use liner time to find the orders with the prices
the user is looking for since it does not know which orders have which
prices, whereas the tree structure is already ordered with cost in mind
since it has to insert the price in the correct spot in the tree to keep
the tree balance.