Add Remove Complated Orders From Orders Event
Add and remove the completed orders from the Orders event to not clog up
the display with finished orders since this will slow the screen.
Furthermore, pagination or infinite scroll will not work in this context
as the page is continually updated.

Instead, we should limit what is being displayed and only show the
active orders as the users will more likely see what is happening. For
example, we can store all the orders in other data structures and search
there, speeding up the search and having a place where we can display an
archive to the user of what orders they have done for the day.
Furthermore, we can make reports from them using this data structure.