Add Message Componet to App Body
Add the Message component to App Body as there are different messages
depending on which mode the body is in search or order list.
For example, if it is in the order list, we should display a no message
that tells the user that they should wait for the orders to come.
Likewise, if the body is in search mode, we should display that no
orders are found in the context searches, not the list.