Add Real Time Update to Search Order Search Tree
Add real-time update to search by Order Search Tree so that the user can
see the latest version of the order, as the Order Search Tree does not
refresh the screen when it gets an update as it happens in the
background. While the orders in state update the screen but are hard to
search through by combining the two, we get the best of both worlds on
does display while the other can search and now update with the latest
version of the state.